@@ -2,21 +2,19 @@ package com.didi.virtualapk.demo.fragment;
 
 import android.content.res.Configuration;
 import android.os.Bundle;
+import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
-import android.util.Log;
-import android.view.LayoutInflater;
 import android.view.View;
-import android.view.ViewGroup;
 import android.widget.ImageView;
 
 import com.didi.virtualapk.demo.R;
+import com.kaolafm.base.BaseFragment;
 
 import butterknife.BindView;
-import butterknife.ButterKnife;
-import me.yokeyword.fragmentation.SupportFragment;
 
-public class LauncherFragment extends SupportFragment {
-    @BindView(R.id.iv1)
+
+public class LauncherFragment extends BaseFragment {
+    @BindView(R.id.iv1x)
     ImageView imageView;
 
     public static LauncherFragment getInstance() {
@@ -29,18 +27,15 @@ public class LauncherFragment extends SupportFragment {
         super.onCreate(savedInstanceState);
     }
 
-    @Nullable
     @Override
-    public View onCreateView(LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
-        View view = inflater.inflate(R.layout.fragment_launcher, container, false);
-        ButterKnife.bind(this, view);
-        imageView.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                extraTransaction().start(new SecondFragment());
-            }
-        });
-        return view;
+    public void onViewCreated(@NonNull View view, @Nullable Bundle savedInstanceState) {
+        super.onViewCreated(view, savedInstanceState);
+        imageView.setOnClickListener(v -> extraTransaction().start(new SecondFragment()));
+    }
+
+    @Override
+    protected int getLayoutId() {
+        return R.layout.fragment_launcher;
     }
 
     @Override
